修复新伤害计算流程

+++ src/main/java/cn/nukkit/item/enchantment/protection/EnchantmentProtectionFall.java
@@ -25,18 +25,18 @@ public class EnchantmentProtectionFall extends EnchantmentProtection {
     }
 
     @Override
-    public double getTypeModifier() {
+    public int getTypeModifier() {
         return 2;
     }
 
     @Override
-    public float getProtectionFactor(EntityDamageEvent e) {
+    public int getProtectionFactor(EntityDamageEvent e) {
         DamageCause cause = e.getCause();
 
         if (level <= 0 || cause != DamageCause.FALL && cause != DamageCause.STALAGMITE) {
             return 0;
         }
 
-        return (float) (getLevel() * getTypeModifier());
+        return getLevel() * getTypeModifier();
     }
 }
